在单线程模式下，禁用CachedAlign，节省内存
