+++ agent/Cargo.toml
@@ -19,6 +19,5 @@ crossbeam-channel = "0.5.1"
 log = "0.4.14"
 ds = { path = "../ds" }
 
-tokio = {version = "1.10.0", features = ["rt", "net", "rt-multi-thread", "time", "macros"]}
-
+tokio = {version = "1.12.0", features = ["rt", "net", "rt-multi-thread", "time", "macros"]}
 mimalloc = { version = "*", default-features = false }
