+++ context/Cargo.toml
@@ -11,8 +11,8 @@ edition = "2018"
 
 clap = "3.0.0-beta.2"
 url = "2.2.2"
-tokio = { version = "1.11.0", features = ["fs"] }
+tokio = { version = "1.12.0", features = ["fs"] }
 log = "0.4.14"
 
 [features]
-listen_all = []
+listen-all = []
