+++ discovery/Cargo.toml
@@ -17,7 +17,7 @@ enum_dispatch = "0.3.7"
 
 #for http request and json parse
 reqwest = {version = "0.11.4", features = ["json"]}
-tokio = { version = "1.11.0", features = ["fs"] }
+tokio = "1.12.0"
 json = "0.12.4"
 
 #for deserialize
