+++ ds/Cargo.toml
@@ -13,3 +13,7 @@ rand = "0.8.4"
 cache_line_size = "1.0.0"
 log = "0.4.14"
 crossbeam-queue = "0.3.2"
+
+[features]
+default = ["single-thread"]
+single-thread = []
