+++ stream/Cargo.toml
@@ -20,7 +20,6 @@ tokio = "1.12.0"
 rand = "0.8.4"
 enum_dispatch = "0.3.7"
 bytes = "1.0.1"
-cache_line_size = "1.0.0"
 byteorder = "1.4.3"
 crossbeam-channel = "0.5.1"
 atomic-waker = "1"
